Move the `check_disk_quota` fn to a new module
